bidouille

+++ balls.h
@@ -1,12 +1,13 @@
 // message philosophique: saute dans une flaque avant que la flaque te saute
 #include <vector>
-#include "tools.h"
+#include "tools.cc"
 
-bool verif_balls(int& x, int& y, int& r, int& dx, int& dy);
+bool verif_balls(double& x, double& y, double& r, double& dx, double& dy);
 
 class Ball : public Cercle {
 public: 
     using Cercle::intersects;
+
     Ball(double x, double y, double r, double dx, double dy)
     : Cercle(x, y, r, dx, dy) {}
  //ajouter un compteur pour transmettre les index au fonction de collisions
